Implemented domain logic GatheringProviderDetails
This is the very domain and the logic of how financial relationships
get created or updated, and thus evaluating a status for the data
onboarding process. It does not contain any technology, involving
user-interface or infrastructure needs. It currently only focuses on
the 'Gathering Provider Details' status.

Non-obvious things:
--
[Testing Approach]
The testing approach might not be the most obvious, but we tried to
bring how the ACs are broken down into the code itself to reduce
back-and-forth understanding between business and engineers.

Because the status 'Gathering Provider Details' is evaluated
differently between client/provider specific financial relationships,
to increase readability and cohesion, we created a different files of
tests for each group. The groups we found are the following:
- all relationships.
- provider-specific relationships.
- client-specific relationships.

As we worked bottom-up, these 3 unique files of tests form a certain
specification. So it made sense for us to create one, which gathers all
of those tests defined in each separate file. We called that
specification, `GatheringProviderDetailsTestSpec`. This makes us able to
test that single specification against the service, which implements
the business logic. So the first thing we see in the tests source code,
is the specification, and under that package, we se all individual
tests that are grouped as mentioned above.

The specification and thus all individual test files are abstractions
that must, at least, depend on the very system under test. We tried our
best to define the tests in such a way that do not depend on
implementation details, and that can be hopefully run against a
different morph of the system under test. Because we currently follow
the Clean Architecture and Ports and Adapters architectural patterns,
where our business logic is exposed via abstractions, and not concrete
implementations; we can use that to our advantage, and create a
different morph of the system under test (polymorphism).

To exercise the specification against that system under test, we need to
implement it and give it whatever is needed for that test to run. Our
current single test is for `UpsertClientFinancialRelationshipUseCase`
against an in-memory version of the driven port (repository), that is
sufficient to pass all tests. An example of the above things mentioned,
another test could be that of `UpsertClientFinancialRelationshipUseCase`
but against an in-memory or dockerised mongodb, for when that adapter is
implemented.

[Testing Library]
We brought in the library called `hamcrest` which uses the junit api,
but offers a lot more flexibility to express assertions with less noise,
and keeping a better domain-specific language. Doing so, it brings the
`Matcher<T>` api, for which there already a couple of high-level
helpers, but we can create our own matchers.